backend/requirement.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
openai>=1.0
python-dotenv>=1.0
qdrant-client>=1.7
sentence-transformers>=2.2
PyPDF2>=3.0
python-docx>=1.0
langchain>=0.1
langchain-qdrant>=0.1
langchain-classic>=0.1